Add files via upload
The folder names match up with the assembly vulnerability, files ending with the X32 or X64 we're used in the following byte format, essentially they were compiled differently with the gcc parameter of m-32, more c code with vulnerabilities will be published on the security branch in the future. All the code has been compiled with gcc and debugged with gdb.